+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed progress speed not updating when total doesn't change
 - Fixed superfluous new line in Status https://github.com/willmcgugan/rich/issues/1662
 - Fixed infinite loop in set_cell_size https://github.com/willmcgugan/rich/issues/1682
+- Fixed superfluous new line in Progress https://github.com/willmcgugan/rich/issues/1691
 
 ### Added